thehyve London hackathon merge with galaxy and other updates

+++ ddl/postgres/tm_wz/wt_subject_microarray_calcs.sql
@@ -4,9 +4,9 @@
 CREATE TABLE wt_subject_microarray_calcs (
     trial_name character varying(50),
     probeset_id bigint,
-    mean_intensity bigint,
-    median_intensity bigint,
-    stddev_intensity bigint
+    mean_intensity numeric,
+    median_intensity numeric,
+    stddev_intensity numeric
 );
 
 --
